requirements.txt
Flask
line-bot-sdk
numpy
openai
Pillow
tensorflow
firebase-admin
google-cloud-firestore
python-dotenv